requirements.txt
python-telegram-bot==21.7
aiohttp==3.11.11
py-clob-client>=0.34.0
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
